+++ ProcessOptimizer/optimizer/optimizer.py
@@ -753,7 +753,7 @@ class Optimizer(object):
                 prob_stbr = 0.25
 
                 # Simulate a random number
-                random_uniform_number = np.random.uniform()
+                random_uniform_number = self.rng.uniform()
 
                 # The random number decides what strategy to use for the next point
                 if random_uniform_number < prob_stbr:
@@ -1309,12 +1309,18 @@ class Optimizer(object):
     # This function calls NSGAII to estimate the Pareto Front
     def NSGAII(self, MU=40):
         from ._NSGA2 import NSGAII
-
+        
+        if MU % 4 != 0:
+            raise ValueError(
+                "Number of simulated points must be divisible by 4 for the NSGAII algorithm"
+            )
+            
         pop, logbook, front = NSGAII(
             self.n_objectives,
             self.__ObjectiveGP,
             np.array(self.space.transformed_bounds),
             MU=MU,
+            seed=42,
         )
 
         return pop, logbook, front