+++ ProcessOptimizer/tests/test_multiobjective.py
@@ -6,6 +6,7 @@ from numpy.testing import assert_equal
 
 
 from ProcessOptimizer import Optimizer
+from ProcessOptimizer.model_systems import get_model_system
 
 
 @pytest.mark.fast_test
@@ -46,3 +47,47 @@ def test_Pareto_in_space():
     # Assert that Pareto points are in space
     for x in pop:
         assert_equal(opt.space.__contains__(x), True)
+
+
+@pytest.mark.fast_test
+def test_Pareto_reproducible():
+    gold_model_system = get_model_system('gold_map')
+    distance_model_system = get_model_system('distance_map', camp_coordinates=(4,10))
+
+    opt = Optimizer(gold_model_system.space, n_initial_points=4, n_objectives=2)
+
+    gold_model_system.noise_model.set_seed(40)
+    distance_model_system.noise_model.set_seed(40)
+
+    for i in range(10):
+        new_dig_site = opt.ask()
+        gold_found = gold_model_system.get_score(new_dig_site)
+        distance = distance_model_system.get_score(new_dig_site)
+        opt.tell(new_dig_site, [gold_found, distance])
+    
+    # Calculate Pareto front twice and compare
+    pop1, logbook, front1 = opt.NSGAII()
+    pop2, logbook, front2 = opt.NSGAII()
+    assert_equal(pop1, pop2)
+    assert_equal(front1, front2)
+
+
+@pytest.mark.fast_test
+def test_multiobjective_ask_reproducible():
+    gold_model_system = get_model_system('gold_map')
+    distance_model_system = get_model_system('distance_map', camp_coordinates=(4,10))
+
+    opt1 = Optimizer(gold_model_system.space, n_initial_points=4, n_objectives=2)
+    opt2 = Optimizer(gold_model_system.space, n_initial_points=4, n_objectives=2)
+
+    gold_model_system.noise_model.set_seed(40)
+    distance_model_system.noise_model.set_seed(40)
+    # Check that the two optimizers stay in sync
+    for i in range(40):
+        new_dig_site_1 = opt1.ask()
+        new_dig_site_2 = opt2.ask()
+        assert_equal(new_dig_site_1, new_dig_site_2)
+        gold_found = gold_model_system.get_score(new_dig_site_1)
+        distance = distance_model_system.get_score(new_dig_site_1)
+        opt1.tell(new_dig_site_1, [gold_found, distance])
+        opt2.tell(new_dig_site_2, [gold_found, distance])